backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.3.0
alembic==1.16.4
httpx==0.28.1
pytest==8.4.2
pytest-cov==6.2.1
